backend/requirements.txt
fastapi
uvicorn
boto3
python-multipart
boto3 
python-dotenv